+++ Runtime/Events/ColliderScriptableEvent.cs
@@ -7,7 +7,7 @@ namespace ScriptableEvents.Events
         menuName = ScriptableEventConstants.MenuNameBase + "/Collider Scriptable Event",
         order = ScriptableEventConstants.MenuOrderUnityObjectEvent + 1
     )]
-    public class ColliderScriptableEvent : BaseScriptableEvent<Collider>
+    public sealed class ColliderScriptableEvent : ScriptableEvent<Collider>
     {
     }
 }