+++ Runtime/Listeners/BoolScriptableEventListener.cs
@@ -6,7 +6,7 @@ namespace ScriptableEvents.Listeners
         ScriptableEventConstants.MenuNameBase + "/Bool Scriptable Event Listener",
         ScriptableEventConstants.MenuOrderPrimitiveEvent + 0
     )]
-    public class BoolScriptableEventListener : BaseScriptableEventListener<bool>
+    public sealed class BoolScriptableEventListener : ScriptableEventListener<bool>
     {
     }
 }