+++ Runtime/Listeners/ColorScriptableEventListener.cs
@@ -6,7 +6,7 @@ namespace ScriptableEvents.Listeners
         ScriptableEventConstants.MenuNameBase + "/Color Scriptable Event Listener",
         ScriptableEventConstants.MenuOrderUnityPrimitiveEvent + 5
     )]
-    public class ColorScriptableEventListener : BaseScriptableEventListener<Color>
+    public sealed class ColorScriptableEventListener : ScriptableEventListener<Color>
     {
     }
 }