+++ Runtime/Listeners/DoubleScriptableEventListener.cs
@@ -6,7 +6,7 @@ namespace ScriptableEvents.Listeners
         ScriptableEventConstants.MenuNameBase + "/Double Scriptable Event Listener",
         ScriptableEventConstants.MenuOrderPrimitiveEvent + 4
     )]
-    public class DoubleScriptableEventListener : BaseScriptableEventListener<double>
+    public sealed class DoubleScriptableEventListener : ScriptableEventListener<double>
     {
     }
 }