+++ Runtime/Listeners/GameObjectScriptableEventListener.cs
@@ -6,8 +6,7 @@ namespace ScriptableEvents.Listeners
         ScriptableEventConstants.MenuNameBase + "/Game Object Scriptable Event Listener",
         ScriptableEventConstants.MenuOrderUnityObjectEvent + 2
     )]
-    public class GameObjectScriptableEventListener
-        : BaseScriptableEventListener<GameObject>
+    public sealed class GameObjectScriptableEventListener : ScriptableEventListener<GameObject>
     {
     }
 }