+++ Runtime/Listeners/IntScriptableEventListener.cs
@@ -6,7 +6,7 @@ namespace ScriptableEvents.Listeners
         ScriptableEventConstants.MenuNameBase + "/Int Scriptable Event Listener",
         ScriptableEventConstants.MenuOrderPrimitiveEvent + 1
     )]
-    public class IntScriptableEventListener : BaseScriptableEventListener<int>
+    public sealed class IntScriptableEventListener : ScriptableEventListener<int>
     {
     }
 }